Refactor profile page for modularity.
