+++ client/src/pages/PlacesPage.jsx
@@ -1,104 +1,22 @@
-import { useState } from "react";
-import { Link, useParams } from "react-router-dom";
-import Perks from "../components/Perks";
-import axios from "axios";
-import PhotosUploader from "../components/PhotosUploader";
+import { Link} from "react-router-dom";
+import AccountNav from "../components/AccountNav";
 
 export default function PlacesPage () {
-    // Grabs domain of current subpage
-    const {action} = useParams();
-    // State for the form inputs
-    const[title, setTitle] = useState('');
-    const[address, setAddress] = useState('');
-    const[addedPhotos, setAddedPhotos] = useState([]); // Array of added photos
-    const[description, setDescription] = useState('');
-    const[extraInfo, setExtraInfo] = useState('');
-    const[perks, setPerks] = useState([]);
-    const[checkIn, setCheckIn] = useState('');
-    const[checkOut, setCheckOut] = useState('');
-    const[maxGuests, setMaxGuests] = useState(1);
-
-    // Functions to simplify code
-    function inputHeader(text) {
-        return (
-            <h2 className="text-2xl mt-4">{text}</h2>
-        );
-    }
-    function inputDescription(text) {
-        return(
-            <p className="text-gray-500 text-sm">{text}</p>
-        );
-    }
-    function preInput(header, description) {
-        return(
-            <>
-                {inputHeader(header)}
-                {inputDescription(description)}
-            </>
-        );
-    }
+    // // Grabs domain of current subpage
+    // const {action} = useParams();
     
-
     return (
         <>
+            <AccountNav/>
             {/* Only displays add new place button if you aren't on the new subpage */}
-            {action !== 'new' && (
-                <div className="text-center">
-                    <Link className="inline-flex gap-1 bg-primary text-white py-2 px-6 rounded-full" to={'/account/places/new'}>
-                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="size-6">
-                            <path fillRule="evenodd" d="M12 3.75a.75.75 0 0 1 .75.75v6.75h6.75a.75.75 0 0 1 0 1.5h-6.75v6.75a.75.75 0 0 1-1.5 0v-6.75H4.5a.75.75 0 0 1 0-1.5h6.75V4.5a.75.75 0 0 1 .75-.75Z" clipRule="evenodd" />
-                        </svg>
-                        Add new place
-                    </Link>
-                </div>
-            )}
-            {action === 'new' && (
-                <div>
-                    <form>
-                        {preInput('Title', 'Title for your place. Should be short and catchy.')}
-                        <input type="text" value={title} onChange={ev => setTitle(ev.target.value)} placeholder="title, for example: My lovely apartment"/>
-                        {preInput('Address', 'Address of this place.')}
-                        <input type="text" value={address} onChange={ev => setAddress(ev.target.value)} placeholder="address"/>
-                        {preInput('Photos', 'More is better.')}
-                        {/* For upload photos by link */}
-                        <PhotosUploader addedPhotos={addedPhotos} onChange={setAddedPhotos}/>
-                        <h2 className="text-2xl mt-4">Description</h2>
-                        <p className="text-gray-500 text-sm">Description of the place.</p>
-                        <textarea value={description} onChange={ev => setDescription(ev.target.value)}/>
-                        {/* Checkbox list of all possible perks. */}
-                        {preInput('Perks', 'Select all the perks of your place.')}
-                        <div className="grid mt-2 gap-2 grid-cols-2 md:grid-cols-3 lg:grid-cols-6">
-                            {/* Component containg all perks selected perks are passed to component */}
-                            <Perks selected={perks} onChange={setPerks}/>
-                        </div>
-                        {preInput('Extra info', 'House rules, etc.')}
-                        <textarea value={extraInfo} onChange={ev => setExtraInfo(ev.target.value)} />
-                        {preInput('Check in & out times', 'Add check in and check out times, remember to leave time for cleaning between guests.')}
-                        <div className="grid gap-2 sm:grid-cols-3">
-                            <div>
-                                <h3 className="mt-2 -mb-1">Check in time</h3>
-                                <input type="text" 
-                                    value={checkIn} 
-                                    onChange={ev => setCheckIn(ev.target.value)} 
-                                    placeholder="14"/>
-                            </div>
-                            <div>
-                                <h3 className="mt-2 -mb-1">Check out time</h3>
-                                <input type="text" 
-                                    value={checkOut} 
-                                    onChange={ev => setCheckOut(ev.target.value)} 
-                                    placeholder="11"/>
-                            </div>
-                            <div>
-                                <h3 className="mt-2 -mb-1">Max number of guests</h3>
-                                <input type="number" value={maxGuests} 
-                                    onChange={ev => setMaxGuests(ev.target.value)}/>
-                            </div>
-                        </div>
-                        <button className="primary my-4">Save</button>
-                    </form>
-                </div>
-            )}
+            <div className="text-center">
+                <Link className="inline-flex gap-1 bg-primary text-white py-2 px-6 rounded-full" to={'/account/places/new'}>
+                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="size-6">
+                        <path fillRule="evenodd" d="M12 3.75a.75.75 0 0 1 .75.75v6.75h6.75a.75.75 0 0 1 0 1.5h-6.75v6.75a.75.75 0 0 1-1.5 0v-6.75H4.5a.75.75 0 0 1 0-1.5h6.75V4.5a.75.75 0 0 1 .75-.75Z" clipRule="evenodd" />
+                    </svg>
+                    Add new place
+                </Link>
+            </div>
         </>
     )
 }
